캐릭터 본체 밸런스 맞춤 아직 저장은 안댐

+++ Assets/Scripts/SaveManager.cs
@@ -101,7 +101,7 @@ public class SaveManager : MonoBehaviour {
 
         public bool[] GFExist = new bool[6];
 
-        public ulong JugallumLev;
+        public int JugallumLev;
         public ulong TouchMoney;
 
         public int noretry;
